+++ src/VaultSharp/Backends/System/Models/SealStatus.cs
@@ -70,5 +70,14 @@ namespace VaultSharp.Backends.System.Models
         /// </value>
         [JsonProperty("cluster_id")]
         public string ClusterId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the nonce.
+        /// </summary>
+        /// <value>
+        /// The nonce.
+        /// </value>
+        [JsonProperty("nonce")]
+        public string Nonce { get; set; }
     }
 }